Begin implementing resources, rename files
Additions:
* `@/app/components/head` for components for page metadata such as the
  title
* `@/app/components/resource-content` for the content of Resources pages

SEO improvements:
* Renamed some directories for user-friendly URLs:
  * /creators/contactForm -> "contact"
  * /privacy&policy -> "privacy-policy"
  * /terms&condition -> "terms-and-conditions"
* Added title in every page

Testimonials component:
* Minor code cleanups

Footer:
* Updated links
* Minor style changes

Brands 7th:
* Added an `id` attribute so that a link in resource 1 can jump to the
  form

Creators 1st, 3rd, last:
* Updated links to contact form

Privacy policy, T&C:
* Fixed title

Resources list:
* Updated titles and path names
* Removed entry for post with no content provided

Resource 1:
* Fixed style errors
* Used new components
