+++ src/app/components/Footer.jsx
@@ -59,14 +59,14 @@ const Footer = ({ hoverText }) => {
               <li
                 className={`hover:text-[#8A5AFE] duration-150 cursor-pointer`}
               >
-                FAQs
+                FAQ
               </li>
             </Link>
             <Link href={"/contact"}>
               <li
                 className={`hover:text-[#8A5AFE] duration-150 cursor-pointer`}
               >
-                Contact
+                Contact Us
               </li>
             </Link>
             <Link href={"/resources"}>
@@ -88,14 +88,14 @@ const Footer = ({ hoverText }) => {
         <div>
           <h1 className="font-bold text-2xl -tracking-[.1rem]">Support</h1>
           <ul className="font-light text-sm">
-            <Link href={"/privacy&policy"}>
+            <Link href={"/privacy-policy"}>
               <li
                 className={`hover:text-[#8A5AFE] duration-150 cursor-pointer`}
               >
                 Privacy Policy
               </li>
             </Link>
-            <Link href={"/terms&condition"}>
+            <Link href={"/terms-and-conditions"}>
               <li
                 className={`hover:text-[#8A5AFE] duration-150 cursor-pointer`}
               >
