+++ src/app/contants/resources.json
@@ -1,37 +1,32 @@
 [
     {
-        "title": "A Starter’s Guide for Content Creators In Malaysia",
+        "title": "Unlocking the Power of Brand Deals: A Guide for Content Creators",
         "image": "https://storage.googleapis.com/landing-cultcreative/resources/Blog%201%20-%20A%20Starter%E2%80%99s%20Guide%20for%20Content%20Creators%20In%20Malaysia.jpg",
-        "path": "/resources/blog1"
+        "path": "/resources/unlocking-the-power-of-brand-deals"
     },
     {
-        "title": "Unlocking the Power of UGC Brand Deals",
+        "title": "Unlocking the Power of Influencer Brand Deals",
         "image": "https://storage.googleapis.com/landing-cultcreative/resources/Blog%202%20-%20Unlocking%20the%20Power%20of%20UGC%20Brand%20Deals.jpg",
-        "path": "/resources/blog2"
+        "path": "/resources/unlocking-the-power-of-influencer-brand-deals"
     },
     {
         "title": "Turn Your Content Creation Passion into Profit",
         "image": "https://storage.googleapis.com/landing-cultcreative/resources/Blog%203%20-%20Turn%20Your%20Content%20Creation%20Passion%20into%20Profit.jpg",
-        "path": "/resources/blog3"
+        "path": "/resources/turn-your-content-creation-passion-into-profit"
     },
     {
         "title": "Key UGC Content Creation Ideas And How To Stay Inspired",
         "image": "https://storage.googleapis.com/landing-cultcreative/resources/Blog%204%20-%20Key%20UGC%20Content%20Creation%20Ideas%20And%20How%20To%20Stay%20Inspired.jpg",
-        "path": "/resources/blog4"
+        "path": "/resources/key-ugc-content-creation-ideas-and-how-to-stay-inspired"
     },
     {
-        "title": "Unlocking UGC Creativity How-To Guide + Examples",
+        "title": "Unlocking UGC Creativity: How-To Guide + Examples",
         "image": "https://storage.googleapis.com/landing-cultcreative/resources/Blog%205%20-%20Unlocking%20UGC%20Creativity%20How-To%20Guide%20%2B%20Examples.jpg",
-        "path": "/resources/blog5"
+        "path": "/resources/unlocking-ugc-creativity"
     },
     {
-        "title": "Why And How Does Creator Marketing Work",
+        "title": "Why and How Does Creator Marketing Work?",
         "image": "https://storage.googleapis.com/landing-cultcreative/resources/Blog%206%20-%20Why%20And%20How%20Does%20Creator%20Marketing%20Work.jpg",
-        "path": "/resources/blog6"
-    },
-    {
-        "title": "The Impact of Malaysian Creator Marketing on Brand Awareness and Sales",
-        "image": "https://storage.googleapis.com/landing-cultcreative/resources/Blog%207%20-%20The%20Impact%20of%20Malaysian%20Creator%20Marketing%20on%20Brand%20Awareness%20and%20Sales.jpg",
-        "path": "/resources/blog7"
+        "path": "/resources/why-and-how-does-creator-marketing-work"
     }
 ]
