+++ src/app/resources/unlocking-the-power-of-brand-deals/page.jsx
@@ -1,0 +1,73 @@
+"use client";
+import * as rc from "@/app/components/resource-content";
+
+const Resource1 = ({ params: { name } }) => {
+  const title = "Unlocking the Power of Brand Deals: A Guide for Content Creators";
+
+  return (
+    <rc.Page title={title}>
+      <rc.Content>
+        <rc.h1>{title}</rc.h1>
+        <rc.p>
+          Content creators have become a pivotal part of social media marketing and entertainment. With the creator economy booming, they now have a higher chance at earning money and popularity through brand deals and partnership, bringing both the creators and businesses more visibility, credibility, and financial success.
+        </rc.p>
+        <rc.p>
+          But navigating the world of content creation isn't always easy and securing brand deals is even harder, so here's a guide on how you can land your next brand deal.
+        </rc.p>
+
+        <rc.h2>Taking on Brand Deals</rc.h2>
+        <rc.p>
+          There’s nothing sweeter than earning your first brand deal. You’ll get to monetise your content (yay, more money!), collaborate with your dream brands, and even reach larger audiences. By following these tips, you can make the most out of your future brand deals.
+        </rc.p>
+        <rc.ol>
+          <rc.li>
+            Authenticity Is Key: Your audience trusts you for your authenticity, so it's important to collaborate with brands that align with your values and that resonate with your audience.
+          </rc.li>
+          <rc.li>
+            Transparency in Disclosure: No one likes a liar and no one likes to get sued either. Always be transparent about your content by adhering to relevant advertising guidelines and clearly disclosing when a content is sponsored.
+          </rc.li>
+          <rc.li>
+            Quality over Quantity: Instead of pursuing numerous brand deals, prioritise the ones that will allow you to produce engaging and authentic content that adds value to both the brand and your audience.
+          </rc.li>
+          <rc.li>
+            Negotiation and Contracts: Familiarise yourself with negotiation tactics and remember to always navigate contracts with caution. Make sure the terms of the contract aligns with your brand and audience, and don't hesitate to seek legal guidance when necessary.
+          </rc.li>
+          <rc.li>
+            Consistent Brand Alignment: Maintain a consistent brand image and messaging across all your brand deals. This ensures cohesive and authentic representation of your brand and values.
+          </rc.li>
+        </rc.ol>
+        <rc.p>
+          Stick to this guide when approaching brand deals, so you can establish yourself as a credible and sought-after content creator whilst also maintaining the trust of your audience.
+        </rc.p>
+
+        <rc.h2>How do you get brand deals on Cult Creative?</rc.h2>
+        <rc.p>
+          Scoring your first (or second or third) brand deal just got easier thanks to Cult Creative’s creator platform. Whether you’re into fashion and beauty, tech and finance, or maybe you’re an explorer who’s ready to seize any opportunity that comes your way, we’ll connect you to amazing brands across South East Asia when you become a CC creator.
+        </rc.p>
+        <rc.p>
+          Did I mention it’s free to use? Yup, it is!
+        </rc.p>
+        <rc.p>
+          Simply fill out{" "}
+          <rc.a href="/brands#form">this form</rc.a>{" "}
+          and you’re on your way to securing your next brand deal.
+        </rc.p>
+
+        <rc.h2>Conclusion</rc.h2>
+        <rc.p>
+          It can be both exciting and complex to work with brands. It is the path to elevating your career and creative endeavours.
+        </rc.p>
+        <rc.p>
+          With this guide, you’ll be well-equipped to forge successful partnerships with brands whilst also retaining the trust and loyalty of your audience. Remember to always be authentic, transparent, and genuine about your content no matter who you’re creating it for.
+        </rc.p>
+      </rc.Content>
+
+      <rc.Keywords>
+        <rc.Keyword value="Creator Economy Malaysia" />
+        <rc.Keyword value="Brand Deals for Creators" />
+      </rc.Keywords>
+    </rc.Page>
+  );
+}
+
+export default Resource1;
